datebase_tool/pyproject.toml
[project]
name = "datebase-tool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pymysql>=1.1.1",
    "python-dotenv>=1.0.1",
    "sqlalchemy>=2.0.39",
]
